Restructure columns into inheritence heirarchy
The various types of columns such as mesure etc. now
have their own classes. This is to provide support for visitors.

RolePlaying Dimensions are removed, and replaced by a DimensionReference,
which acts like a columns AND a dimension, and deals with the naming
issue (i.e. the reference and the underlying dimension do not have
to have the same name).