Allow pivoting columns in Chicago::Query.
The pivoting syntax is the following:

  sales.total.sum ~ sales.customer.boolean_flag

Which, in this case will add 2 columns to the dataset,
one for where the flag is true, one for where is is false.

Pivoting is not generally supported at the moment - it
is only currently supported in the case where the possible
elements for a column's values can be determined from a
column defintion, and the set is "reasonably" sized. This
means Booleans, Strings with :elements defined and Integers
which have a total range of size 500 or less.

Pivoting can only be done at one level, and can only sensibly
be executed on calculated columns.

Current sequel support is now limited up to 3.29.0