Change the interface of Query to take an AST.
Instead of dealing with query fragments as strings, Query now takes an
AST. If a string-based query language is requiredin future, a separate
parser will be built for it.

Also separates out interacting with the generated dataset from adding
columns, filters etc. to the query.

These changes should make it easier to have queries as JSON objects.